Re-wrote to accept a new argument: --symbols.
Symbols is a list of open/close pairs. Now, the checker scans characters
for a match against the dictionary of open/close pairs, and keeps
a dictionary of stacks. Also added doctests and removed the list
reduction stuff into a new function (since it's much more complicated
now anyway, as it has to flatten a whole dictionary structure)